# docs/openapi.yaml
openapi: 3.1.0
info:
  title: api2.autisme.lu
  version: 1.0.0
  description: Modern Node.js API for NamNam and legacy Autisme Luxembourg workflows.
servers:
  - url: https://api2.autisme.lu
security:
  - bearerAuth: []
paths:
...
  /people:
    get:
      summary: List people
      parameters:
        - { name: active, in: query, schema: { type: boolean } }
        - { name: meal_registration_enabled, in: query, schema: { type: boolean } }
        - { name: is_encadrant, in: query, schema: { type: boolean } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200': { description: List of people }
    post:
      summary: Create person
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PersonInput' }
      responses:
        '201': { description: Created person }
  /people/{id}:
    get:
      summary: Get person
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Person }
    patch:
      summary: Update person
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Updated person }
  /people/{id}/toggle:
    patch:
      summary: Toggle person boolean field
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Updated person }
  /people/{id}/billing-category:
    patch:
      summary: Update billing category
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Updated person }
...
  /base-plans:
    get:
      summary: Base plan by person
      parameters:
        - { name: person_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200': { description: Base plan entries }
    post:
      summary: Upsert base plan entries
      responses:
        '201': { description: Saved entries }
...
  /base-plans/missing:
    get:
      summary: Missing meal registrations for date
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Missing entries }
  /meal-registrations:
    get:
      summary: List meal registrations
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: until, in: query, schema: { type: string, format: date } }
        - { name: person_id, in: query, schema: { type: integer } }
        - { name: place_id, in: query, schema: { type: integer } }
        - { name: meal_choice_id, in: query, schema: { type: integer } }
      responses:
        '200': { description: Registrations }
  /meal-registrations/by-person-date:
    get:
      summary: One registration and warnings
      parameters:
        - { name: person_id, in: query, required: true, schema: { type: integer } }
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Registration and warnings }
...
  /weekly-menus:
    get:
      summary: Get weekly menu
      parameters:
        - { name: monday_date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Weekly menu }
    post:
      summary: Upsert weekly menu
      responses:
        '201': { description: Saved weekly menu }
...
  /menu-assets:
    get:
      summary: List menu assets
      responses:
        '200': { description: Assets }
    post:
      summary: Upload menu asset
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                kind: { type: string }
                altText: { type: string }
      responses:
        '201': { description: Uploaded asset }
...
  /billing/data:
    get:
      summary: Billing data
      parameters:
        - { name: month, in: query, required: true, schema: { type: string, pattern: '^\\d{4}-\\d{2}$' } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        '200': { description: Billing rows }
...
  /billing/non-billable-dates/{date}:
    delete:
      summary: Delete non-billable date
      parameters:
        - { name: date, in: path, required: true, schema: { type: string, format: date } }
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    PersonInput:
      type: object
      properties:
        lastName: { type: string }
        firstName: { type: string }
        isEncadrant: { type: boolean }
        mealRegistrationEnabled: { type: boolean }
        billingCategory: { type: string }